// src/models/_db.ts
export interface _IIssue {
    id: number;
    courseId: number;
    accountId: number;
    title: string;
    openAt: number;
    closerId?: number;
    closeAt?: number;
    status: number;

}

export interface _IIssueComment {
    id: number;
    issueId: number;
    accountId: number;
    createAt: number;
    content: string;
    status: number;

}